Update JPAFundingDAO for test

+++ sharebook/src/main/java/com/sharebook/sharebook/dao/JPAFundingDAO.java
@@ -19,26 +19,51 @@ public class JPAFundingDAO implements FundingDAO {
 
 	@Override
 	public void create(Funding funding) throws DataAccessException {
+		if (em == null) {
+			System.out.println("==== EntityManager is null ====");
+			return;
+		}
+		
 		em.persist(funding);
 	}
 
 	@Override
 	public void delete(int funding_id) throws DataAccessException {
+		if (em == null) {
+			System.out.println("==== EntityManager is null ====");
+			return;
+		}
+		
 		em.remove(findFundingById(funding_id));
 	}
 
 	@Override
 	public void update(int funding_id, Funding funding) throws DataAccessException {
+		if (em == null) {
+			System.out.println("==== EntityManager is null ====");
+			return;
+		}
+		
 		em.merge(funding);
 	}
 
 	@Override
 	public Funding findFundingById(int funding_id) throws DataAccessException {
+		if (em == null) {
+			System.out.println("==== EntityManager is null ====");
+			return null;
+		}
+		
 		return em.find(Funding.class, funding_id);
 	}
 
 	@Override
 	public List<Funding> findFundingByTitle(String title) throws DataAccessException {
+		if (em == null) {
+			System.out.println("==== EntityManager is null ====");
+			return null;
+		}
+		
 		TypedQuery<Funding> query = em.createQuery(
 				"select funding from FUNDING funding "
 				+ "where funding.title=?1", Funding.class);
@@ -48,6 +73,11 @@ public class JPAFundingDAO implements FundingDAO {
 
 	@Override
 	public List<Funding> findFundingByAuthor(String author) throws DataAccessException {
+		if (em == null) {
+			System.out.println("==== EntityManager is null ====");
+			return null;
+		}
+		
 		TypedQuery<Funding> query = em.createQuery(
 				"select funding from FUNDING funding "
 				+ "where funding.author=?1", Funding.class);
@@ -57,6 +87,11 @@ public class JPAFundingDAO implements FundingDAO {
 
 	@Override
 	public List<Funding> findFundingList() throws DataAccessException {
+		if (em == null) {
+			System.out.println("==== EntityManager is null ====");
+			return null;
+		}
+		
 		TypedQuery<Funding> query = em.createQuery(
 				"select funding from FUNDING funding ", Funding.class);
 		return query.getResultList();
@@ -64,6 +99,11 @@ public class JPAFundingDAO implements FundingDAO {
 
 	@Override
 	public List<Funding> findFundingList(int user_id) throws DataAccessException {
+		if (em == null) {
+			System.out.println("==== EntityManager is null ====");
+			return null;
+		}
+		
 		TypedQuery<Funding> query = em.createQuery(
 				"select funding from FUNDING funding "
 				+ "where funding.user_id=?1", Funding.class);
